requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
langgraph==0.2.18  # Use the latest version as of July 2025; check PyPI for updates
google-api-python-client==2.97.0
python-dotenv==1.0.0
requests==2.31.0